# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "iris-daemon",
    "description": "",
    "license": {
      "name": "GPL-3.0-or-later",
      "identifier": "GPL-3.0-or-later"
    },
    "version": "0.1.0-beta"
  },
  "paths": {
...
    "/api/v1/projects/{id}": {
      "get": {
        "tags": [],
        "operationId": "project",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Project": {
        "type": "object",
        "required": [
          "id",
          "root",
          "name",
          "created_at",
          "last_opened_at",
          "hidden",
          "cache_root",
          "groups_dirty",
          "pending_analysis"
        ],
        "properties": {
          "cache_root": {
            "type": "string"
          },
          "created_at": {
            "type": "string"
          },
          "groups_dirty": {
            "type": "boolean",
            "description": "Group membership/ranking must be rebuilt, including after reopening a project."
          },
          "hidden": {
            "type": "boolean"
          },
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "last_opened_at": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "pending_analysis": {
            "type": "integer",
            "format": "int64",
            "description": "Active photos without current-engine analysis, independent of pagination.",
            "minimum": 0
          },
          "root": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "session": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "security": [
    {
      "session": []
    }
  ]
}